// src/model/common.ts
/**
 * @description 倍速基础类型参数
 */
export interface OptionVal {
    label: string
    value: number | string
    default?: boolean
}
/**
 * @description 水印参数
 */
export interface WaterMaskOption {
    show: boolean
    url?: string
    style?: string
}

export interface VideoOption {

}